[Dy2St] Cleanup duplicate dy2st `test_simnet_v2` and `test_ptb_lm_v2`
